Binary storage : check if list of file is not nul when deleting a directory

+++ meveo-admin/ejbs/src/main/java/org/meveo/service/storage/FileSystemService.java
@@ -85,9 +85,14 @@ public class FileSystemService {
 		}
 
 		String storage = getStoragePath(params, values);
-		File dir = new File(storage);
-		for(File file : dir.listFiles()) {
-			file.delete();
+		File directory = new File(storage);
+		if(directory.exists() && directory.isDirectory()) {
+			final File[] files = directory.listFiles();
+			if(files != null) {
+				for (File file : files) {
+					file.delete();
+				}
+			}
 		}
 	}
